+++ src/dashInternoTela.js
@@ -1,10 +1,15 @@
 import { useContext } from "react";
 import VtInternaCards from "./components/VtInternaCards";
+import RegistroVotos from "./components/ListRegistroVotos/index";
 import styles from "./dashInternoTela.module.css";
 import { Dropdown } from "primereact/dropdown";
 import "primereact/resources/themes/saga-blue/theme.css";
 import "primereact/resources/primereact.min.css";
 import { TurmaContext } from "./context/TurmaContext";
+import ContainerChart from "./components/ContainerChart";
+import ChartBar from "./components/ChartBar";
+import ChartDonut from "./components/ChartDonut";
+import ChartSemanal from "./components/ChartSemanal";
 
 export default function DashInternoMain() {
   const {
@@ -15,6 +20,11 @@ export default function DashInternoMain() {
     setSelectedCurso,
     setSelectedVotacaoType,
   } = useContext(TurmaContext);
+
+  const chartData = [
+    { name: "Votos Confirmados", value: 50 },
+    { name: "Não Votaram", value: 50 },
+  ];
 
   return (
     <>
@@ -56,9 +66,30 @@ export default function DashInternoMain() {
 
       {/* Fora da div vermelha, abaixo normalmente */}
       <>
-      <div className={styles.cardContent}>
-        <VtInternaCards />
-      </div>
+        <div className={styles.cardContent}>
+          <VtInternaCards
+            conteudo={
+              <div>
+                <h2 className={styles.titleChart}>Overview</h2>
+                <div className={styles.chartContainer}>
+                  <ContainerChart props={<ChartBar />} />
+                  <ContainerChart
+                    props={
+                      <ChartDonut
+                        title="Quantidade de Votos"
+                        data={chartData}
+                      />
+                    }
+                  />
+                  <ContainerChart props={<ChartSemanal />} />
+                </div>
+                <div className={styles.listContainer}>
+                  <RegistroVotos />
+                </div>
+              </div>
+            }
+          />
+        </div>
       </>
     </>
   );
